refactor: move perf test to own file
