+++ apps/api/src/test/index.js
@@ -17,7 +17,7 @@ concurrently(
             name: 'api setup',
             env: { DATABASE_URL, NODE_ENV: 'test', PORT: 4001 },
         },
-        { command: 'npx jest --config jest.json', name: 'jest', env: { DATABASE_URL } },
+        { command: 'npx jest --config jest.json --runInBand', name: 'jest', env: { DATABASE_URL } },
     ],
     {
         killOthers: ['failure', 'success'],
